Move the BasicSearchStrategy to init and add documentation
